Fix safety-critical bugs: crashes, UB, and data corruption
- Add missing return after null file error in sendSysexCartridge (crash)
- Mask algorithm index to 0-31 range in dx7note init/update (OOB crash)
- Change sysex offset from uint8 to int to prevent truncation (wrong memory write)
- Guard PluginFx::process against zero-length buffers (crash in some DAWs)
- Add default case to filter switch to prevent uninitialized variable
- Clamp shift amounts in Exp2::lookup and Freqlut::lookup (undefined behavior)
- Null-terminate opSwitch after strncpy (buffer overread)
- Add null check for InputStream in setupStartupCart (null deref)
- Remove memcpy that overwrote normalized sysex data (dead code / data corruption)
- Clamp portamento_cc loaded from XML to 0-127 (OOB array access)
- Guard pitchStep division against zero and out-of-range values (division by zero)

+++ Source/PluginProcessor.cpp
@@ -714,7 +714,7 @@ void DexedAudioProcessor::handleIncomingMidiMessage(MidiInput* source, const Mid
             return;
             }
 
-            uint8 offset = (buf[3] << 7) + buf[4];
+            int offset = (buf[3] << 7) + buf[4];
             uint8 value = buf[5];
 
             TRACE("parameter change message offset:%d value:%d", offset, value);
